LayoutTests/fast/forms/input-widths-expected.txt

Text field types should have the same default widths.

PASS document.getElementById("email").offsetWidth is baseWidth
PASS document.getElementById("number").offsetWidth is baseWidth
PASS document.getElementById("password").offsetWidth is baseWidth
PASS document.getElementById("tel").offsetWidth is baseWidth
PASS document.getElementById("url").offsetWidth is baseWidth
PASS document.getElementById("date").offsetWidth is baseWidth
PASS document.getElementById("datetime").offsetWidth is baseWidth
PASS document.getElementById("datetimelocal").offsetWidth is baseWidth
PASS document.getElementById("month").offsetWidth is baseWidth
PASS document.getElementById("time").offsetWidth is baseWidth
PASS document.getElementById("week").offsetWidth is baseWidth
PASS successfullyParsed is true

TEST COMPLETE
